// Frontend/merchant-admin/src/types/billing.ts
export interface PaystackTransactionInitialize {
  authorization_url: string;
  access_code: string;
  reference: string;
}

export interface PaystackTransactionVerify {
  id: number;
  domain: string;
  status: string;
  reference: string;
  amount: number;
  message: string | null;
  gateway_response: string;
  paid_at: string;
  created_at: string;
  channel: string;
  currency: string;
  ip_address: string;
    metadata: any;
  customer: {
    id: number;
    first_name: string | null;
    last_name: string | null;
    email: string;
    customer_code: string;
    phone: string | null;
        metadata: any;
    risk_action: string;
  };
  authorization: {
    authorization_code: string;
    bin: string;
    last4: string;
    exp_month: string;
    exp_year: string;
    channel: string;
    card_type: string;
    bank: string;
    country_code: string;
    brand: string;
    reusable: boolean;
    signature: string;
  };
}

export interface PaystackAccountResolve {
  account_number: string;
  account_name: string;
  bank_id: number;
}